changes in player and coin + added a new class
removed the coin and made a sound when Mario takes it, in player.
In the coin class i am trying to give the player points when he takes
the coin.
added a score class where i am trying to get the score from when the
player take a coin on the screen.